move all single state simulation duplicates to distinct plugin
