+++ bundles/org.palladiosimulator.analyzer.slingshot.singlestatesimulation.application/src/org/palladiosimulator/analyzer/slingshot/singlestatesimulation/application/SingeStateSimulationApplication.java
@@ -14,11 +14,11 @@ import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.emf.ecore.resource.impl.ResourceSetImpl;
 import org.eclipse.equinox.app.IApplication;
 import org.eclipse.equinox.app.IApplicationContext;
+import org.palladiosimulator.analyzer.slingshot.singlestatesimulation.workflow.SingleStateSimulationRootJob;
+import org.palladiosimulator.analyzer.slingshot.singlestatesimulation.workflow.SingleStateSimulationWorkflowConfiguration;
 import org.palladiosimulator.analyzer.slingshot.stateexploration.explorer.ui.ExplorationConfiguration;
 import org.palladiosimulator.analyzer.slingshot.stateexploration.workflow.ExplorationWorkflowConfiguration;
-import org.palladiosimulator.analyzer.slingshot.stateexploration.workflow.SingleStateSimulationWorkflowConfiguration;
 import org.palladiosimulator.analyzer.slingshot.stateexploration.workflow.jobs.ExplorationRootJob;
-import org.palladiosimulator.analyzer.slingshot.stateexploration.workflow.jobs.SingleStateSimulationRootJob;
 import org.palladiosimulator.experimentautomation.application.ExperimentApplication;
 import org.palladiosimulator.experimentautomation.application.tooladapter.abstractsimulation.AbstractSimulationConfigFactory;
 import org.palladiosimulator.experimentautomation.application.tooladapter.stateexploration.model.StateExplorationConfiguration;
