+++ bundles/org.palladiosimulator.analyzer.slingshot.singlestatesimulation/META-INF/MANIFEST.MF
@@ -1,0 +1,28 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Singlestatesimulation
+Bundle-SymbolicName: org.palladiosimulator.analyzer.slingshot.singlestatesimulation
+Bundle-Version: 1.0.0.qualifier
+Automatic-Module-Name: org.palladiosimulator.analyzer.slingshot.singlestatesimulation
+Bundle-RequiredExecutionEnvironment: JavaSE-17
+Export-Package: org.palladiosimulator.analyzer.slingshot.singlestatesimulation.workflow
+Require-Bundle: de.uka.ipd.sdq.workflow,
+ de.uka.ipd.sdq.workflow.mdsd.blackboard,
+ org.palladiosimulator.analyzer.slingshot.core,
+ org.palladiosimulator.analyzer.slingshot.stateexploration.workflow,
+ org.palladiosimulator.analyzer.workflow,
+ org.palladiosimulator.analyzer.slingshot.networking.data,
+ org.palladiosimulator.analyzer.slingshot.stateexploration,
+ org.palladiosimulator.pcm,
+ org.palladiosimulator.mdsdprofiles.api,
+ org.palladiosimulator.analyzer.slingshot.stateexploration.data,
+ de.uka.ipd.sdq.simucomframework,
+ org.palladiosimulator.analyzer.slingshot.snapshot.data,
+ org.palladiosimulator.analyzer.slingshot.workflow,
+ org.palladiosimulator.spd,
+ org.palladiosimulator.analyzer.slingshot.stateexploration.converter,
+ org.palladiosimulator.analyzer.slingshot.behavior.spd.data,
+ org.palladiosimulator.monitorrepository,
+ org.palladiosimulator.servicelevelobjective,
+ org.palladiosimulator.analyzer.slingshot.common.utils,
+ org.palladiosimulator.analyzer.slingshot.workflow.events
