+++ bundles/org.palladiosimulator.analyzer.slingshot.singlestatesimulation/build.properties
@@ -1,0 +1,4 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .
