+++ bundles/org.palladiosimulator.analyzer.slingshot.singlestatesimulation/src/org/palladiosimulator/analyzer/slingshot/singlestatesimulation/workflow/SingleStateSimulationWorkflowConfiguration.java
@@ -1,10 +1,11 @@
-package org.palladiosimulator.analyzer.slingshot.stateexploration.workflow;
+package org.palladiosimulator.analyzer.slingshot.singlestatesimulation.workflow;
 
 
 import java.nio.file.Path;
 import java.util.Map;
 
 import org.palladiosimulator.analyzer.slingshot.core.api.SimulationConfiguration;
+import org.palladiosimulator.analyzer.slingshot.stateexploration.workflow.ExplorationWorkflowConfiguration;
 import org.palladiosimulator.analyzer.slingshot.workflow.events.PCMWorkflowConfiguration;
 
 import de.uka.ipd.sdq.simucomframework.SimuComConfig;
